style: adding resources.go file
This PR adds new functions to upload OCI images and handle upload
requests. In the initial commit, all these functions were
added to the `applications.go` file which reduces the readability of code.
This commit puts all the functions which are used to process
OCI image resources in a separate file named `resources.go'
to reduce the complexity and increase the readilibity.

docs(application): Application documentation is updated by adding
custom resource usage in the existing sample TF plan.
Application optional TF plan parameters are updated
by changing the description of `resources` map.

test(resources): A few tests are added for the functions in
the resources.go. This tests will be expanded in the next commits.
