+++ internal/juju/applications_test.go
@@ -19,7 +19,6 @@ import (
 	"github.com/juju/names/v4"
 	"github.com/juju/utils/v3"
 	"github.com/juju/version/v2"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/mock/gomock"
 )
@@ -388,20 +387,4 @@ func (s *ApplicationSuite) TestReadApplicationRetryNotFoundStorageNotFoundError(
 // a normal test function and pass our suite to suite.Run
 func TestApplicationSuite(t *testing.T) {
 	suite.Run(t, new(ApplicationSuite))
-}
-
-func TestNewEndpointPath(t *testing.T) {
-	application := "ausf"
-	name := "sdcore-ausf-k8s"
-	want := "/applications/ausf/resources/sdcore-ausf-k8s"
-	got := newEndpointPath(application, name)
-	assert.Equal(t, got, want)
-}
-
-func TestNewEndpointPathEmptyInputs(t *testing.T) {
-	application := ""
-	name := ""
-	want := "/applications//resources/"
-	got := newEndpointPath(application, name)
-	assert.Equal(t, got, want)
 }
