+++ main.go
@@ -36,7 +36,7 @@ var (
 func main() {
 	var debugMode bool
 
-	flag.BoolVar(&debugMode, "debug", false, "set to true to run the provider with support for debuggers like delve")
+	flag.BoolVar(&debugMode, "debug", true, "set to true to run the provider with support for debuggers like delve")
 	flag.Parse()
 
 	var serveOpts []tf6server.ServeOpt
